Added ipUtils file
